added google signin, password reset

+++ src/components/RegisterComponent.jsx
@@ -1,31 +1,49 @@
 import React, { useState } from "react";
-import { RegisterAPI } from "../api/AuthAPI";
+import { RegisterAPI, GoogleSignInAPI } from "../api/AuthAPI";
 import { postUserData } from "../api/FirestoreAPI";
 import LinkedinLogo from "../assets/linkedinLogo.png";
 import { useNavigate } from "react-router-dom";
 import { getUniqueID } from "../helpers/getUniqueId";
 import { toast } from "react-toastify";
+import GoogleSignIn from "./GoogleComponent";
 
 export default function RegisterComponent() {
   let navigate = useNavigate();
   const [credentails, setCredentials] = useState({});
   const register = async () => {
     try {
-      let res = await RegisterAPI(credentails.email, credentails.password);
-      toast.success("Account Created!");
-      postUserData({
-        userID: getUniqueID(),
-        name: credentails.name,
-        email: credentails.email,
-        imageLink:
-          "https://t4.ftcdn.net/jpg/02/29/75/83/240_F_229758328_7x8jwCwjtBMmC6rgFzLFhZoEpLobB6L8.jpg",
-      });
-      navigate("/home");
+      if (!credentails.email || !credentails.password || !credentails.name) {
+        toast.error("Please enter name, email, and password");
+        return;
+      }
+      const res = await RegisterAPI(credentails.email, credentails.password);
+      console.log(res.code);
+      if (res.code === "auth/invalid-email") {
+        toast.error("invalid email");
+        return;
+      }
+      if (res.code === "auth/weak-password") {
+        toast.error("weak password");
+        return;
+      }
+      if (res.code === "auth/email-already-in-use") {
+        toast.error("user already exists. Login");
+        return;
+      }
       localStorage.setItem("userEmail", res.user.email);
     } catch (err) {
       console.log(err);
       toast.error("Cannot Create your Account");
     }
+    toast.success("Signed In to Linkedin!");
+    postUserData({
+      userID: getUniqueID(),
+      name: credentails.name,
+      email: credentails.email,
+      imageLink:
+        "https://t4.ftcdn.net/jpg/02/29/75/83/240_F_229758328_7x8jwCwjtBMmC6rgFzLFhZoEpLobB6L8.jpg",
+    });
+    navigate("/home");
   };
 
   return (
@@ -70,7 +88,7 @@ export default function RegisterComponent() {
               className="w-full p-3 text-gray-700 border rounded focus:outline-gray-900 focus-within:shadow-lg"
             />
           </div>
-          <div className="py-6 text-center">
+          <div className="py-4 text-center">
             <button
               onClick={register}
               className="w-full py-2 font-bold text-white bg-blue-700 rounded-full hover:bg-blue-800 focus:outline-none focus:shadow-outline"
@@ -78,8 +96,8 @@ export default function RegisterComponent() {
               Agree & Join
             </button>
           </div>
-          <p className="text-center">OR</p>
-          {/* TODO: Google */}
+          <p className="py-4 text-center ">OR</p>
+          <GoogleSignIn />
           <div className="my-4 ">
             <p className="text-center ">
               Already on LinkedIn? &nbsp;
